Cut dependencies of LLMPromptBlueprint to higher levels
