+++ pipelex/tools/misc/context_provider_abstract.py
@@ -1,0 +1,20 @@
+from abc import ABC, abstractmethod
+from typing import Any, Dict, Optional, Type
+
+from pipelex.tools.exceptions import ToolException
+
+
+class ContextProviderException(ToolException):
+    def __init__(self, variable_name: str, message: str, *args: object, **kwargs: object) -> None:
+        self.variable_name = variable_name
+        super().__init__(message, *args, **kwargs)
+
+
+class ContextProviderAbstract(ABC):
+    @abstractmethod
+    def get_typed_object_or_attribute(self, name: str, wanted_type: Optional[Type[Any]] = None) -> Any:
+        pass
+
+    @abstractmethod
+    def generate_context(self) -> Dict[str, Any]:
+        pass
